# docs/1_openapi.yaml
openapi: 3.0.3
info:
  title: AuctionHub API
  description: REST API for the AuctionHub real-time auction platform
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateCheckoutSessionRequest:
      type: object
      required: [auctionId]
      properties:
        auctionId:
          type: integer
          format: int64
          example: 42
    CheckoutSessionResponse:
      type: object
      properties:
        sessionId:
          type: string
          example: cs_test_a1b2c3d4e5f6g7h8i9j0
        checkoutUrl:
          type: string
          example: https://checkout.stripe.com/pay/cs_test_a1b2c3d4e5f6g7h8i9j0
        paymentId:
          type: integer
          format: int64
          example: 25
    PaymentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 25
        auctionId:
          type: integer
          format: int64
          example: 42
        auctionTitle:
          type: string
          example: Vintage Rolex Submariner 1965
        buyerId:
          type: integer
          format: int64
          example: 13
        buyerName:
          type: string
          example: alice_dev
        sellerId:
          type: integer
          format: int64
          example: 7
        sellerName:
          type: string
          example: bob_collector
        amount:
          type: number
          example: 5250.00
        status:
          type: string
          enum: [PENDING, COMPLETED, FAILED, REFUNDED, CANCELLED]
          example: COMPLETED
        paidAt:
          type: string
          format: date-time
          example: "2026-03-03T11:05:22"
        createdAt:
          type: string
          format: date-time
          example: "2026-03-03T11:00:00"
...
    ErrorResponse:
      type: object
      required: [code, message, error, timestamp]
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: "Auction not found with id: 42"
        error:
          type: string
          example: Bad Request
        timestamp:
          type: string
          format: date-time
          example: "2026-03-02T14:30:00"

# ── Reusable responses ────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 401
            message: Missing or invalid authentication token
            error: Unauthorized
            timestamp: "2026-03-02T14:30:00"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 500
            message: An unexpected error occurred
            error: Internal Server Error
            timestamp: "2026-03-02T14:30:00"
paths:
...
  /payments/create-checkout-session:
    post:
      tags: [Payments]
      summary: Create Stripe checkout session (winner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCheckoutSessionRequest'
            example:
              auctionId: 42
      responses:
        '200':
          description: Checkout session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutSessionResponse'
              example:
                sessionId: cs_test_a1b2c3d4e5f6g7h8i9j0
                checkoutUrl: https://checkout.stripe.com/pay/cs_test_a1b2c3d4e5f6g7h8i9j0
                paymentId: 25
        '400':
          description: Auction not ended or payment already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: A payment already exists for this auction
                error: Bad Request
                timestamp: "2026-03-03T11:00:00"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Not the auction winner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 403
                message: Only the auction winner can make a payment
                error: Forbidden
                timestamp: "2026-03-03T11:00:00"
        '404':
          description: Auction or user not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "Auction not found with id: 42"
                error: Not Found
                timestamp: "2026-03-03T11:00:00"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /payments/webhook:
    post:
      tags: [Payments]
      summary: Stripe webhook endpoint (public — no JWT required)
      security: []
      parameters:
        - name: Stripe-Signature
          in: header
          required: true
          schema:
            type: string
            example: t=1700000000,v1=abc123signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
      responses:
        '200':
          description: Webhook processed
          content:
            application/json:
              example: Webhook received
        '400':
          description: Invalid Stripe signature or bad payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 400
                message: Invalid Stripe webhook signature
                error: Bad Request
                timestamp: "2026-03-03T11:00:00"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /payments/session/{sessionId}:
    get:
      tags: [Payments]
      summary: Get payment by Stripe session ID
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            example: cs_test_a1b2c3d4e5f6g7h8i9j0
      responses:
        '200':
          description: Payment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
              example:
                id: 25
                auctionId: 42
                auctionTitle: Vintage Rolex Submariner 1965
                buyerName: alice_dev
                sellerName: bob_collector
                amount: 5250.00
                status: COMPLETED
                paidAt: "2026-03-03T11:05:22"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "Payment not found for session: cs_test_a1b2c3d4e5f6g7h8i9j0"
                error: Not Found
                timestamp: "2026-03-03T11:00:00"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /payments/auction/{auctionId}:
    get:
      tags: [Payments]
      summary: Get payment by auction ID
      parameters:
        - name: auctionId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 42
      responses:
        '200':
          description: Payment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
              example:
                id: 25
                auctionId: 42
                auctionTitle: Vintage Rolex Submariner 1965
                buyerName: alice_dev
                sellerName: bob_collector
                amount: 5250.00
                status: COMPLETED
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Auction or payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "No payment found for auction: 42"
                error: Not Found
                timestamp: "2026-03-03T11:00:00"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /payments/my-payments:
    get:
      tags: [Payments]
      summary: Get authenticated user's payments as buyer
      responses:
        '200':
          description: Payment history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentResponse'
              example:
                - id: 25
                  auctionId: 42
                  auctionTitle: Vintage Rolex Submariner 1965
                  buyerName: alice_dev
                  sellerName: bob_collector
                  amount: 5250.00
                  status: COMPLETED
                  paidAt: "2026-03-03T11:05:22"
                - id: 18
                  auctionId: 35
                  auctionTitle: Sony A7III Camera Body
                  buyerName: alice_dev
                  sellerName: emma_gear
                  amount: 1800.00
                  status: COMPLETED
                  paidAt: "2026-02-20T15:30:00"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "User not found: alice_dev"
                error: Not Found
                timestamp: "2026-03-03T11:00:00"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /payments/payments-received:
    get:
      tags: [Payments]
      summary: Get payments received as seller
      responses:
        '200':
          description: Received payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentResponse'
              example:
                - id: 31
                  auctionId: 48
                  auctionTitle: Gibson Les Paul 1959 Reissue
                  buyerName: emma_watch
                  sellerName: alice_dev
                  amount: 3200.00
                  status: COMPLETED
                  paidAt: "2026-03-01T09:45:00"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 404
                message: "User not found: alice_dev"
                error: Not Found
                timestamp: "2026-03-01T09:30:00"
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ── Reviews ──
...